@@ -1,4 +1,4 @@
-class Project < ActiveRecord::Base
+class Project < ApplicationRecord
     validates :name, length: {minimum: 3}, uniqueness: true
     has_many :reviews, dependent: :destroy
     belongs_to :user
